--- minimal-build/README.md
@@ -1,5 +1,0 @@
-```sh
-pnpm dev
-pnpm build
-pnpm preview
-```
