--- minimal-build/package.json
@@ -1,21 +1,0 @@
-{
-	"type": "module",
-	"scripts": {
-		"dev": "vite",
-		"build": "vite build",
-		"preview": "vite preview",
-		"lint": "biome check --apply --linter-enabled=false ."
-	},
-	"dependencies": {
-		"react": "19.0.0-rc-38e3b23483-20240529",
-		"react-dom": "19.0.0-rc-38e3b23483-20240529",
-		"react-server-dom-webpack": "19.0.0-rc-38e3b23483-20240529"
-	},
-	"devDependencies": {
-		"@types/node": "^20.12.13",
-		"@types/react": "^18.3.3",
-		"@types/react-dom": "^18.3.0",
-		"vite": "^5.2.12"
-	},
-	"packageManager": "pnpm@8.15.8+sha512.d1a029e1a447ad90bc96cd58b0fad486d2993d531856396f7babf2d83eb1823bb83c5a3d0fc18f675b2d10321d49eb161fece36fe8134aa5823ecd215feed392"
-}
