--- minimal-build/tsconfig.json
@@ -1,14 +1,0 @@
-{
-	"include": ["src", "vite.config.ts"],
-	"compilerOptions": {
-		"verbatimModuleSyntax": true,
-		"strict": true,
-		"noEmit": true,
-		"moduleResolution": "Bundler",
-		"module": "ESNext",
-		"target": "ESNext",
-		"lib": ["ESNext", "DOM", "DOM.Iterable"],
-		"types": ["vite/client"],
-		"jsx": "react-jsx"
-	}
-}
