+++ minimal/README.md
@@ -1,5 +1,3 @@
-RSC SSR and client component demo (single process + two vite servers)
-
 ```sh
 pnpm dev
 pnpm build
