+++ minimal/tsconfig.json
@@ -8,6 +8,7 @@
 		"module": "ESNext",
 		"target": "ESNext",
 		"lib": ["ESNext", "DOM", "DOM.Iterable"],
+		"types": ["vite/client"],
 		"jsx": "react-jsx"
 	}
 }
