+++ minimal/vite.config.ts
@@ -1,4 +1,13 @@
-import { type InlineConfig, build, createServer, defineConfig } from "vite";
+import fs from "node:fs";
+import path from "node:path";
+import {
+	type InlineConfig,
+	type Manifest,
+	type Plugin,
+	build,
+	createServer,
+	defineConfig,
+} from "vite";
 import { $__global } from "./src/global";
 
 export default defineConfig((_env) => ({
@@ -19,7 +28,45 @@ export default defineConfig((_env) => ({
 					});
 				};
 			},
+			async configurePreviewServer(server) {
+				const mod = await import(path.resolve("dist/ssr/index.js"));
+				return () => {
+					server.middlewares.use(async (req, res, next) => {
+						try {
+							await mod.default(req, res);
+						} catch (e) {
+							next(e);
+						}
+					});
+				};
+			},
 		},
+		// virtual modules to switch dev/build behavior
+		createVirtualPlugin("client-references", () => {
+			// build only
+			const ids = [...$__global.clientReferences];
+			return `export default { ${ids
+				.map((id) => `"${id}": () => import("${id}"),\n`)
+				.join("")} }`;
+		}),
+		createVirtualPlugin("ssr-assets", () => {
+			let ssrAssets: any;
+			if ($__global.reactServer) {
+				// dev
+				ssrAssets = {
+					bootstrapModules: ["/@vite/client", "/src/entry-browser"],
+				};
+			} else {
+				// build
+				const manifest: Manifest = JSON.parse(
+					fs.readFileSync("dist/browser/.vite/manifest.json", "utf-8"),
+				);
+				ssrAssets = {
+					bootstrapModules: [manifest["src/entry-browser.tsx"].file],
+				};
+			}
+			return `export default ${JSON.stringify(ssrAssets)}`;
+		}),
 		// setup/teardown 2nd vite server from main vite server
 		{
 			name: "react-server:dev",
@@ -60,12 +107,16 @@ export default defineConfig((_env) => ({
 	],
 	// browser build config
 	build: {
+		manifest: true,
 		outDir: "dist/browser",
 		rollupOptions: {
 			input: {
 				index: "/src/entry-browser",
 			},
 		},
+	},
+	optimizeDeps: {
+		include: [],
 	},
 }));
 
@@ -104,8 +155,9 @@ const reactServerViteConfig: InlineConfig = {
 		{
 			name: "client-reference",
 			transform(code, id, _options) {
-				$__global.clientReferences.delete(id);
 				// client reference transform
+				// (in practice, it's critical to post-process `id` to match how Vite handles them on browser)
+				$__global.clientReferences.delete(id);
 				if (/^(("use client")|('use client'))/.test(code)) {
 					$__global.clientReferences.add(id);
 					const matches = code.matchAll(/export function (\w+)\(/g);
@@ -123,3 +175,18 @@ const reactServerViteConfig: InlineConfig = {
 		},
 	],
 };
+
+function createVirtualPlugin(name: string, load: Plugin["load"]) {
+	name = "virtual:" + name;
+	return {
+		name: `virtual-${name}`,
+		resolveId(source, _importer, _options) {
+			return source === name ? "\0" + name : undefined;
+		},
+		load(id, options) {
+			if (id === "\0" + name) {
+				return (load as any)(id, options);
+			}
+		},
+	} satisfies Plugin;
+}
